Extract chat action helpers
