+++ js/chat.js
@@ -19,7 +19,7 @@ import { buildLabContext, getContextSummary, injectLensChunks } from './lab-cont
 import { hasLens, queryLensMulti, updateLensIndicator } from './lens.js';
 import { applyInlineMarkdown, renderMarkdown } from './markdown.js';
 import { renderProfileContextCards } from './context-cards.js';
-import { CHAT_ICON_COPY, CHAT_ICON_REFRESH, setIconButtonContent } from './chat-icons.js';
+import { setIconButtonContent } from './chat-icons.js';
 import {
   closeSummaryModal, copySummary, deleteSavedSummary, downloadSummary,
   printSummary, renderSavedSummaries, summarizeThread, viewSavedSummary,
@@ -45,6 +45,9 @@ import {
 import {
   clearChatHistory, getChatStorageKey, loadChatHistory, saveChatHistory,
 } from './chat-history.js';
+import {
+  buildActionBar, copyMessage, regenerateLastMessage, toggleContextDetails,
+} from './chat-actions.js';
 export {
   autoResizePersonaTextarea, deleteCustomPersonality, editCustomPersonality,
   generateCustomPersonality, getActivePersonality, getCustomPersonalities,
@@ -57,11 +60,18 @@ export {
 export {
   clearChatHistory, getChatStorageKey, loadChatHistory, saveChatHistory,
 } from './chat-history.js';
+export {
+  buildActionBar, copyMessage, regenerateLastMessage, toggleContextDetails,
+} from './chat-actions.js';
 
 // ═══════════════════════════════════════════════
 // ABORT CONTROLLER (stop streaming)
 // ═══════════════════════════════════════════════
 let _chatAbortController = null;
+
+export function isChatStreaming() {
+  return !!_chatAbortController;
+}
 
 // ═══════════════════════════════════════════════
 // TYPEWRITER — smooth character trickle for streaming
@@ -153,81 +163,6 @@ function _updateWebSearchToggleVisibility() {
   if (label) label.style.display = supportsWebSearch() ? '' : 'none';
 }
 export function refreshWebSearchToggle() { _updateWebSearchToggleVisibility(); }
-
-// ═══════════════════════════════════════════════
-// ACTION BAR RENDERING
-// ═══════════════════════════════════════════════
-export function buildActionBar(msgIndex) {
-  const msg = state.chatHistory[msgIndex];
-  if (!msg || msg.role !== 'assistant') return '';
-  const isLast = msgIndex === state.chatHistory.length - 1;
-
-  let html = '<div class="chat-action-bar">';
-  if (isLast) {
-    html += `<button class="chat-action-btn" onclick="regenerateLastMessage()" title="Regenerate response">${CHAT_ICON_REFRESH}<span>Regenerate</span></button>`;
-  }
-  html += `<button class="chat-action-btn" onclick="copyMessage(${msgIndex})" id="chat-copy-btn-${msgIndex}" title="Copy to clipboard">${CHAT_ICON_COPY}<span>Copy</span></button>`;
-  html += '</div>';
-
-  // Context used section
-  if (msg.context && msg.context.length > 0) {
-    html += `<div class="chat-context-toggle" onclick="toggleContextDetails(${msgIndex})">`;
-    html += `<span class="chat-toggle-arrow" id="chat-ctx-arrow-${msgIndex}">\u25B8</span> Context used (${msg.context.length} area${msg.context.length !== 1 ? 's' : ''})`;
-    html += '</div>';
-    html += `<div class="chat-context-details" id="chat-ctx-details-${msgIndex}" style="display:none">`;
-    for (const area of msg.context) {
-      html += `<span class="chat-context-item">\u2713 ${escapeHTML(area.label)}${area.detail ? ' (' + escapeHTML(area.detail) + ')' : ''}</span>`;
-    }
-    html += '</div>';
-  }
-
-  return html;
-}
-
-// ═══════════════════════════════════════════════
-// ACTION HANDLERS
-// ═══════════════════════════════════════════════
-export function regenerateLastMessage() {
-  if (state.chatHistory.length < 2) return;
-  if (_chatAbortController) return; // streaming in progress
-  // Pop the last assistant message
-  state.chatHistory.pop();
-  // Get the last user message
-  const lastUserMsg = state.chatHistory[state.chatHistory.length - 1];
-  if (!lastUserMsg || lastUserMsg.role !== 'user') return;
-  // Re-fill input and re-send
-  const input = document.getElementById('chat-input');
-  if (input) input.value = lastUserMsg.content;
-  // Remove the user message too (sendChatMessage will re-add it)
-  state.chatHistory.pop();
-  saveChatHistory();
-  renderChatMessages();
-  sendChatMessage();
-}
-
-export function copyMessage(msgIndex) {
-  const msg = state.chatHistory[msgIndex];
-  if (!msg) return;
-  const btn = document.getElementById(`chat-copy-btn-${msgIndex}`);
-  if (!navigator.clipboard) { if (btn) { setIconButtonContent(btn, 'x', 'Not supported'); setTimeout(() => { setIconButtonContent(btn, 'copy', 'Copy'); }, 1500); } return; }
-  navigator.clipboard.writeText(msg.content).then(() => {
-    if (btn) {
-      setIconButtonContent(btn, 'check', 'Copied');
-      setTimeout(() => { setIconButtonContent(btn, 'copy', 'Copy'); }, 1500);
-    }
-  }).catch(() => {
-    if (btn) { setIconButtonContent(btn, 'x', 'Failed'); setTimeout(() => { setIconButtonContent(btn, 'copy', 'Copy'); }, 1500); }
-  });
-}
-
-export function toggleContextDetails(msgIndex) {
-  const details = document.getElementById(`chat-ctx-details-${msgIndex}`);
-  const arrow = document.getElementById(`chat-ctx-arrow-${msgIndex}`);
-  if (!details) return;
-  const open = details.style.display !== 'none';
-  details.style.display = open ? 'none' : 'flex';
-  if (arrow) arrow.textContent = open ? '\u25B8' : '\u25BE';
-}
 
 // ═══════════════════════════════════════════════
 // MESSAGE RENDERING
@@ -2070,6 +2005,7 @@ function _resumeAI() {
 
 Object.assign(window, {
   _resumeAI,
+  isChatStreaming,
   toggleChatFullscreen,
   getChatStorageKey,
   getChatThreadsKey,
