+++ service-worker.js
@@ -72,6 +72,7 @@ const APP_SHELL = [
   '/js/export.js',
   '/js/chat.js',
   '/js/chat-attestation.js',
+  '/js/chat-actions.js',
   '/js/chat-icons.js',
   '/js/chat-personalities.js',
   '/js/chat-history.js',
